Move static methods in SourceI to SourceUtil
